move common routesTest helper code
