+++ app/test-utils/routesTestHelper.js
@@ -1,0 +1,30 @@
+import {
+  FakeAuthProvider,
+  fakeSessionManager,
+} from 'buying-catalogue-library';
+import { App } from '../app';
+import { routes } from '../routes';
+
+const mockLogoutMethod = jest.fn().mockImplementation(() => Promise.resolve({}));
+
+const mockAuthorisedJwtPayload = JSON.stringify({
+  id: '88421113',
+  name: 'Cool Dude',
+  ordering: 'manage',
+  primaryOrganisationId: 'org-id',
+});
+
+export const mockAuthorisedCookie = `fakeToken=${mockAuthorisedJwtPayload}`;
+
+const mockUnauthorisedJwtPayload = JSON.stringify({
+  id: '88421113', name: 'Cool Dude',
+});
+
+export const mockUnauthorisedCookie = `fakeToken=${mockUnauthorisedJwtPayload}`;
+
+export const setUpFakeApp = () => {
+  const authProvider = new FakeAuthProvider(mockLogoutMethod);
+  const app = new App(authProvider).createApp();
+  app.use('/', routes(authProvider, fakeSessionManager()));
+  return app;
+};
